test/fixtures/wpt/encoding/sniffing.html

<!doctype html>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<title>No (UTF-8) sniffing allowed</title>
<div id=log></div>
<div id=content> Hello World! </div>
<script>
test(() => {
  assert_equals(document.getElementById("content").textContent, "\u00C2\u20AC\u00C2\u20AC Hello World! \u00C2\u20AC\u00C2\u20AC");
  assert_equals(document.characterSet, "windows-1252");
});
</script>
